Split test view as viewhttp1 and view
